backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
openai-whisper==20231117
python-dotenv==1.0.0
torch
reportlab==4.0.6